+++ lib/properties_intf.ml
@@ -729,6 +729,8 @@ type place_items =
   | Center
   | Stretch
   | Baseline
+  | First_baseline
+  | Last_baseline
   | Start_safe
   | End_safe
   | Center_safe